+++ smokeTest.js
@@ -1,6 +1,6 @@
 const { helpers } = require('brave-alert-lib')
 const axios = require('axios').default
-const { getRandomArbitrary, getRandomInt, printRandomIntArray } = require('./testingHelpers')
+const { getRandomArbitrary, getRandomInt } = require('./testingHelpers')
 const RADAR_TYPE = require('./RadarTypeEnum')
 const XETHRU_STATE = require('./SessionStateXethruEnum')
 
@@ -32,36 +32,6 @@ async function setup(recipientPhoneNumber, twilioPhoneNumber, type) {
       twilioNumber: twilioPhoneNumber,
       radarType: type,
     })
-  } catch (e) {
-    helpers.log(e)
-  }
-}
-
-async function innosentMovement(coreID, min, max) {
-  try {
-    await axios.post('/api/innosent', {
-      name: 'Radar',
-      data: `{ "inPhase": "${printRandomIntArray(min, max, 15)}", "quadrature": "${printRandomIntArray(min, max, 15)}"}`,
-      ttl: 60,
-      published_at: '2021-03-09T19:37:28.176Z',
-      coreid: `${coreID}`,
-    })
-    await helpers.sleep(1000)
-  } catch (e) {
-    helpers.log(e)
-  }
-}
-
-async function innosentSilence(coreID) {
-  try {
-    await axios.post('/api/innosent', {
-      name: 'Radar',
-      data: `{ "inPhase": "${printRandomIntArray(0, 0, 15)}", "quadrature": "${printRandomIntArray(0, 0, 15)}"}`,
-      ttl: 60,
-      published_at: '2021-03-09T19:37:28.176Z',
-      coreid: `${coreID}`,
-    })
-    await helpers.sleep(1000)
   } catch (e) {
     helpers.log(e)
   }
@@ -135,32 +105,6 @@ async function smokeTest(recipientPhoneNumber, twilioPhoneNumber) {
     }
     for (let i = 0; i < 85; i += 1) {
       await xeThruSilence(radar_coreID)
-    }
-  } catch (error) {
-    helpers.log(`Error running smoke test: ${error}`)
-  } finally {
-    await teardown()
-  }
-  try {
-    helpers.log('Running Innosent Smoke Tests')
-    await teardown()
-    await setup(recipientPhoneNumber, twilioPhoneNumber, RADAR_TYPE.INNOSENT)
-    for (let i = 0; i < 15; i += 1) {
-      await innosentMovement(radar_coreID, getRandomInt(MOV_THRESHOLD + 1, 100), getRandomInt(MOV_THRESHOLD + 1, 100))
-    }
-    await im21Door(door_coreID, IM21_DOOR_STATUS.OPEN)
-    for (let i = 0; i < 5; i += 1) {
-      await innosentMovement(radar_coreID, getRandomInt(0, MOV_THRESHOLD), getRandomInt(0, MOV_THRESHOLD))
-    }
-    await im21Door(door_coreID, IM21_DOOR_STATUS.CLOSED)
-    for (let i = 0; i < 15; i += 1) {
-      await innosentMovement(radar_coreID, getRandomInt(0, MOV_THRESHOLD), getRandomInt(0, MOV_THRESHOLD))
-    }
-    for (let i = 0; i < 15; i += 1) {
-      await innosentMovement(radar_coreID, getRandomInt(MOV_THRESHOLD + 1, 100), getRandomInt(MOV_THRESHOLD + 1, 100))
-    }
-    for (let i = 0; i < 85; i += 1) {
-      await innosentSilence(radar_coreID)
     }
   } catch (error) {
     helpers.log(`Error running smoke test: ${error}`)